fix: skipRounding implemented for TE [DHIS2-17105]
